requirements.txt
# HTTP клиент для запросов к Max.ru API
requests==2.31.0

# Загрузка переменных окружения из .env
python-dotenv==1.0.0